feat(RHINENG-6037): move config method to workaround circular import
