+++ api/staleness_query.py
@@ -1,7 +1,7 @@
 from sqlalchemy.orm.exc import NoResultFound
 
-from app import inventory_config
 from app.auth import get_current_identity
+from app.common import inventory_config
 from app.logging import get_logger
 from app.models import Staleness
